Rework our chunk-background gutter renderer
This is probably not 100% correct, and could be improved, but the
rendering now looks to be basically what we want. The only obvious
remaining issue is that it doesn't show chunk colours after an initial
comparison; possibly an issue with us not queueing it for draw on chunk
change or something.