Remove DiffGrid
While capable, this widget has been painful for a while, because it's a
custom implementation of GtkGrid with a bunch of extra display tied in
to Gdk (which is hard to port to GTK 4) and custom size request logic
(which is even harder to port).

While it might not be gone forever, for the purposes of doing the GTK 4
port I'm going to remove this and accept the (IMO relatively manageable)
feature regression, and then hopefully we can add the functionality back
one day.

If I were to look at doing this again, I would probably look at whether
we can use libpanel (as used by GNOME Builder) as a replacement that
also has a few other bits and pieces that we might like.