Rewrite ActionGutter for snapshot drawing
While not strictly necessary for the port, this has also moved away from
using the themed drawing APIs that are going away in favour of
snapshotting a child GtkButton that we just move around. In theory we
could stop doing a bunch of custom handling for button locations and
motion handlers and just use child widgets, but that does seem like
another step and could be slow if we're managing a hundred buttons at
once.