Fix missing drawing of insert line indicators on last chunk of file
Because we were requesting the chunk from the bounds of the last line,
we weren't drawing an insert indicator for the last chunk of the file
because we assign inserts to the line number *after* they appear, and
there is no next line number in the buffer.

The fix here isn't difficult, but it does involve making a few more
buffer location lookups that would be nice to cache instead at some
point.