Clear action gutter current pointer chunk on chunk changes (#935)
The action gutter normally only recalculates the pointer chunk on
motion/enter/leave events, but we need to also do this when the actual
underlying chunk set changes. This is normally an invisible issue due
to small cursor movements, etc. but a double-click or very controlled
clicks without cursor movement easily reproduce the problem.

Technically this is not an ideal solution. We *should* recalculate the
current pointer chunk when the chunks change instead of nulling it out,
but that would require us to keep the pointer coordinates and do a few
other things that I don't really want to do for the GTK 3 branch at
this point.