Update file-save dialog with callback-based flow and no overwrite check
The dialog.run() API is no longer available so we need to make these
interactions all callback-based.

The overwrite prompt that we had implemented manually has actually
already been present in the default file chooser for a while and could
have been removed already, but we'll take this opportunity to get rid
of it.