Move to using AdwApplication
I didn't want to do this as part of the immediate migration, but just
trying to get dialog replacements working is extremely painful without
just using AdwAlertDialog. The built-in GtkAlertDialog is extremely
inflexible and doesn't let me set a CSS class on the Delete button,
apparently because this isn't part of GTK (never mind the destructive
CSS class being defined in GTK, who cares).

Eventually I'll look to migrate more stuff across, but for now this is
*only* for dialog interactions.